Extract the duplicated methods and move to PermissionTemplates module.
