+++ app/models/concerns/hyrax/permissions.rb
@@ -3,5 +3,6 @@ module Hyrax
     extend ActiveSupport::Concern
     include Hyrax::Permissions::Writable
     include Hyrax::Permissions::Readable
+    include Hyrax::Permissions::PermissionTemplates
   end
 end
